Rename UserTracking component to CookieConsentBanner
